+++ src/main/java/com/dfsek/seismic/algorithms/sampler/noise/cellular/CellularStyleSampler.java
@@ -194,7 +194,7 @@ public abstract class CellularStyleSampler extends NoiseFunction {
     protected final double twoDCellularJitter;
     protected final double threeDCellularJitter;
     protected static final long DOUBLE_ARRAY_BASE = UnsafeUtils.DOUBLE_ARRAY_BASE;
-    protected static final int DOUBLE_ARRAY_SHIFT = UnsafeUtils.DOUBLE_ARRAY_SHIFT;
+    protected static final long DOUBLE_ARRAY_SHIFT = UnsafeUtils.DOUBLE_ARRAY_SHIFT;
     public CellularStyleSampler(double frequency, long salt, Sampler noiseLookup, DistanceFunction distanceFunction,
                                 CellularReturnType returnType,
                                 double jitterModifier, boolean saltLookup) {
